pyproject.toml
[tool.poetry]
name = "orgs-backend"
version = "0.1.0"
description = ""
authors = ["Mateus Santos <santos.mateus@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pycardano = "^0.9.0"
python-jose = "^3.3.0"
cryptography = "^41.0.5"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
tortoise-orm = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
